// src/app/kimios-client-api/model/metaFeed.ts
/**
 * Kimios API
 * Kimios Sample API
 *
 * OpenAPI spec version: 1.2.1 - 7fdc88b413cfe49ca0d23a002cafdcc0ee5ee6ab
 * Contact: documentation@kimios.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface MetaFeed { 
    uid?: number;
    name?: string;
    className?: string;
}